Download NBM data

Just a draft of a function that download all NBM data related to CORI works It motsly works in my setup (download in ~/data_swamp)

Usage

dl_nbm(
  path_to_dl = "~/data_swamp",
  release_date = "June 30, 2023",
  data_type = "Fixed Broadband",
  data_category = "Nationwide",
  ...
)

Arguments

path_to_dl

a string by default "~/data_swamp"

release_date

a string can be "December 31, 2023" or "June 30, 2023"

data_type

a string "Fixed Broadband"

data_category

a string "Nationwide"

...

additional parameters for download.file()

Value

A lot of zipped file

Examples

if (FALSE) { # \dontrun{
system("mkdir -p  ~/data_swamp")
dl_nbm(release_date = "June 30, 2023")
} # }