Load part of NBM at Census Block from CORI s3 bucket

Get all the data related to a states or county.

Usage

get_nbm_bl(geoid_co)

Arguments

geoid_co

a string of 5-digit numbers

Value

a data frame

Details

A row in this data represent a census block (2020 vintage). Use get_fcc_dictionary("nbm_block") to get a description of the date.

Data Source: FCC Broadband Data Collection

Examples

nbm_bl <- get_nbm_bl(geoid_co = "47051")